refactor: v0.9.3 — run loop dedup, FrameState, style split, field unification
Extract run_frame() generic over TerminalBackend trait, replacing 3x300-line
duplicated run loops with shared frame function. Bundle 15+ frame-local
variables into FrameState, reducing Context::new() from 17 params.

- refactor: TerminalBackend trait + run_frame<T>() in lib.rs (-208 lines)
- refactor: FrameState struct replaces 15+ loose variables
- refactor: Context::new(&mut FrameState) removes too_many_arguments
- refactor: style/color.rs + style/theme.rs extracted from style.rs (-664 lines)
- refactor: ContainerBuilder bg_color/dark_bg_color → bg/dark_bg
