+++ src/context/widgets_display.rs
@@ -563,8 +563,8 @@ impl Context {
             border: None,
             border_sides: BorderSides::all(),
             border_style: Style::new().fg(border),
-            bg_color: None,
-            dark_bg_color: None,
+            bg: None,
+            dark_bg: None,
             dark_border_style: None,
             group_hover_bg: None,
             group_hover_border_style: None,
