+++ src/terminal.rs
@@ -39,6 +39,13 @@ pub(crate) struct InlineTerminal {
     color_depth: ColorDepth,
 }
 
+pub(crate) trait TerminalBackend {
+    fn size(&self) -> (u32, u32);
+    fn buffer_mut(&mut self) -> &mut Buffer;
+    fn flush(&mut self) -> io::Result<()>;
+    fn handle_resize(&mut self) -> io::Result<()>;
+}
+
 impl Terminal {
     pub fn new(mouse: bool, kitty_keyboard: bool, color_depth: ColorDepth) -> io::Result<Self> {
         let (cols, rows) = terminal::size()?;
@@ -179,6 +186,24 @@ impl Terminal {
             cursor::MoveTo(0, 0)
         )?;
         Ok(())
+    }
+}
+
+impl TerminalBackend for Terminal {
+    fn size(&self) -> (u32, u32) {
+        Terminal::size(self)
+    }
+
+    fn buffer_mut(&mut self) -> &mut Buffer {
+        Terminal::buffer_mut(self)
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        Terminal::flush(self)
+    }
+
+    fn handle_resize(&mut self) -> io::Result<()> {
+        Terminal::handle_resize(self)
     }
 }
 
@@ -328,6 +353,24 @@ impl InlineTerminal {
             cursor::MoveTo(0, 0)
         )?;
         Ok(())
+    }
+}
+
+impl TerminalBackend for InlineTerminal {
+    fn size(&self) -> (u32, u32) {
+        InlineTerminal::size(self)
+    }
+
+    fn buffer_mut(&mut self) -> &mut Buffer {
+        InlineTerminal::buffer_mut(self)
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        InlineTerminal::flush(self)
+    }
+
+    fn handle_resize(&mut self) -> io::Result<()> {
+        InlineTerminal::handle_resize(self)
     }
 }
